requirements.txt
ffmpeg==1.4
firebase_admin==6.5.0
Flask==3.0.3
Flask_Migrate==4.0.7
flask_sqlalchemy==3.1.1
protobuf==5.28.1
Requests==2.32.3
yt_dlp==2024.8.6
psycopg2-binary
flask-jwt-extended
python-dotenv~=1.0.1
werkzeug~=3.0.4
alembic~=1.13.2
SQLAlchemy~=2.0.34